refactor(lists): extract ListsMobileView from ListsView
